% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_interface.R
\name{criar_interface}
\alias{criar_interface}
\title{Função para rodar as análises automaticamente}
\usage{
criar_interface(dados)
}
\arguments{
\item{a}{É um dataframe, que é o mesmo que seu conjunto de dados.}
}
\description{
Esse função executa uma interface gráfica que vc pode selecionar as variáveis
que devem ser comparadas. Ao selecionar as variáveis a interface mostra um
boxplot e o resultados do teste de comparação que pode ser teste t,
Mann-Whitney, Anova ou Kruskal-Wallis. A análise é selecionada automaticamente
a partir do atendimento dos pressupostos.
}
\examples{
criar_interface(iris)

}
