% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analises_execute.R
\name{testar_Ralef}
\alias{testar_Ralef}
\title{Função para rodar as análises automaticamente}
\usage{
testar_Ralef(dados, grupo)
}
\arguments{
\item{a}{É a coluna de dados numéricos.}

\item{b}{É a coluna de dados categóricos.}
}
\description{
Esta função executa um algorítimo que avalia os dados e roda a análise
que melhor se adequa parar eles. Entre as análises estão o teste t,
Mann-Whitney, Anova e Kruskal-Wallis.
}
\examples{
testar_Ralef(iris$Sepal.Width, iris$Species)

}
